+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Allow to hide "Date Created" and "Date Updated" columns
 - Move "actions" button to bottom to reflect changes in new Craft CMS
 - Allow selecting multiple lines with shift key
+- Export button directly in messages list with the ability to select which messages to export
 
 ## 3.1.0 - 2024-01-10